update migration scripts
