+++ quadratic-api/package.json
@@ -25,8 +25,8 @@
     "docker:dev": "npm run docker:up && dotenv -e .env.test -- prisma migrate deploy && dotenv -e .env.test -- ts-node-dev --watch src --project ./tsconfig.json -- ./src/server.ts && npm run docker:down",
     "docker:test:dev": "npm run docker:up && dotenv -e .env.test -- prisma migrate deploy && dotenv -e .env.test -- jest --watch --detectOpenHandles && npm run docker:down",
     "docker:test:ci": "npm run docker:up && dotenv -e .env.test -- prisma migrate reset --force && dotenv -e .env.test -- jest --runInBand && npm run docker:down",
-    "script:migrateFiles": "ts-node-dev --project ./tsconfig.json -- ./scripts/migrateFiles.ts",
-    "script:migrateUsersToTeams": "ts-node-dev --project ./tsconfig.json -- ./scripts/migrateUsersToTeams.ts",
+    "script:migrate-1": "ts-node-dev --project ./tsconfig.json -- ./scripts/migrate-1.ts",
+    "script:migrate-2": "ts-node-dev --project ./tsconfig.json -- ./scripts/migrate-2.ts",
     "key:generate": "openssl rand -hex 32"
   },
   "author": "David Kircos",
